% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\docType{data}
\name{enhancers}
\alias{enhancers}
\title{Enhancer list}
\format{A data frame with 149097 rows and 5 variables:
\describe{
  \item{chr}{chromosome of enhancers}
  \item{start}{start point of enhancers}
  \item{end}{end point of enhancers}
}}
\source{
\url{http://resource.psychencode.org/Datasets/Derived/DER-03a_hg19_PEC_enhancers.bed}
}
\usage{
enhancers
}
\description{
A dataset containing enhancer regions
}
\keyword{datasets}
