% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexpr.R
\docType{data}
\name{gexpr}
\alias{gexpr}
\title{Sample gene expression data}
\format{An object of class \code{data.frame} with 5000 rows and 1800 columns.}
\usage{
gexpr
}
\description{
Each row represents a gene and each column represents an observation.
}
\keyword{datasets}
