% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hic_data.R
\docType{data}
\name{hic_data}
\alias{hic_data}
\title{Hic data}
\format{a data frame with 149097 rows and 5 variables
\describe{
 \item{chr}{chromosome}
 \item{start1}{start point of one interaction region}
 \item{end1}{end point of one interaction region}
 \item{start2}{start point of the other interaction region}
 \item{end2}{end point of the other interaction region}
}}
\source{
hic data can be found in 
((\url{http://resource.psychencode.org/Datasets/Integrative/Promoter-anchored_chromatin_loops.bed}))
}
\usage{
hic_data
}
\description{
Hic data
}
\keyword{datasets}
