% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_data.R
\docType{data}
\name{interaction_data}
\alias{interaction_data}
\title{Sample interaction data}
\format{A data frame with 100 rows and 7 variables:
\describe{
  \item{gene}{gene id}
  \item{gene_chr}{chromosome for genes}
  \item{enh_chr}{chromosome for enhancers}
  \item{promoter_start}{start point of promoters}
  \item{promoter_end}{end point of enhancers}
  \item{enhancer_start}{start point of enhancers}
  \item{enhancer_end}{end point of enhancers}
}}
\usage{
interaction_data
}
\description{
A dataset containing  gene, gene_chr, promoter_start, promoter_end, enh_chr,
enh_start and enh_end
}
\keyword{datasets}
