% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_gexpr.R
\docType{data}
\name{sample_gexpr}
\alias{sample_gexpr}
\title{Sample gene expression data}
\format{a data frame with 100 rows and 7 variables
\describe{
 \item{columns}{each column represents an observation}
 \item{row}{each row represents a gene}
}}
\usage{
sample_gexpr
}
\description{
Sample gene expression data
}
\keyword{datasets}
