% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_interaction.R
\docType{data}
\name{sample_interaction}
\alias{sample_interaction}
\title{Sample interaction data}
\format{a data frame with 100 rows and 7 variables
\describe{
 \item{gene}{here use gene symbol as gene id}
 \item{gene_chr}{chromosome for target gene}
 \item{promoter_start}{start point of promoter}
 \item{promoter_end}{end point of promoter}
 \item{enhancer_start}{start point of enhancer}
 \item{promoter_end}{end point of enhancer}
}}
\source{
hic data, enhancer data and data can be found in 
((\url{http://resource.psychencode.org/Datasets/}))
}
\usage{
sample_interaction
}
\description{
It is a data set generated by function scGRN_interaction
}
\keyword{datasets}
