% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_TFs.R
\docType{data}
\name{sample_TFs}
\alias{sample_TFs}
\title{Sample TFs data}
\format{a data.table with 100 rows and 7 variables
\describe{
 \item{gene}{gene id}
 \item{promoter}{promoter region for target gene}
 \item{enhancer}{linked enhancer for a target gene}
 \item{promoter_TF}{possible transcription factors that would bind the promoter, its element type is list}
 \item{enhancer_TF}{possible transcription factors that bind to the enhancer, ts element type is list}
}}
\usage{
sample_TFs
}
\description{
This data set can directly be generated use the function scGRN_getTF and data set sample_interation
}
\keyword{datasets}
