% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFs.R
\docType{data}
\name{TFs}
\alias{TFs}
\title{Sample TFs}
\format{A data frame with 100 rows and 5 variables:
\describe{
  \item{gene}{gene id}
  \item{promoter}{promoter_region}
  \item{enhancer}{enhancer region}
  \item{promoter_TF}{transcription factors that bind to the promoter region}
  \item{enhancer_TF}{transcription factors that bind to the enhancer region}
}}
\usage{
TFs
}
\description{
A data.table containing gene, promoter, enhancer, promoter_TF, enhancer_TF
}
\keyword{datasets}
